Reorganized `_vispy`,
- refactored `_vispy` directory to have more organized structure
- improved docs
- added few more tests
- tidied up code
