+++ napari_1d/_qt/qt_toolbar.py
@@ -1,5 +1,4 @@
 """Toolbar"""
-# Third-party imports
 from qtpy.QtCore import Qt
 from qtpy.QtWidgets import QWidget
