+++ napari_1d/_vispy/canvas.py
