+++ napari_1d/_vispy/layers/centroids.py
@@ -1,12 +1,12 @@
 """Centroids layer"""
-from typing import TYPE_CHECKING
+import typing as ty
 
 import numpy as np
 from napari._vispy.layers.base import VispyBaseLayer
 from vispy.scene.visuals import Line as LineVisual
 
-if TYPE_CHECKING:
-    from ..layers import Centroids
+if ty.TYPE_CHECKING:
+    from ...layers import Centroids
 
 
 def make_centroids(data: np.ndarray, orientation: str) -> np.ndarray:
