+++ napari_1d/_vispy/overlays/text.py
