+++ scripts/qtaicons.bat
@@ -1,0 +1,4 @@
+:: Start-up QtAwesome icon browser
+call conda activate napari1d
+call qta-browser
+call conda deactivate
